Incorporated Happiness Graph into the Dashboard. Currently uses dates pulled from models.py, but not yet using live happiness percentages.
Same goes for Steps graph - live date values, but not live step counts
yet.

Also - the dates are out of order at the moment.

Added an example Bootstrap progress bar. Should be easy to hook this up
with live values for current_experience and required_experience to get a
progress percentage.